requirements.txt
numpy
lightning
hydra-core
wandb
segmentation-models-pytorch
rasterio